move drill card
